requirements.txt
django-secrets
django-tailwind
pokemontcgsdk